requirements.txt
blinker==1.9.0
click==8.1.8
diskcache==5.6.3
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
llama_cpp_python==0.3.8
MarkupSafe==3.0.2
numpy==2.2.5
typing_extensions==4.13.2
Werkzeug==3.1.3
